Unifying remote and local calls

+++ src/modules/NuFFT.hpp
@@ -71,7 +71,7 @@ namespace RRStrategy {
 		
 	private:
 		
-		NFFT<double> ft;
+		NFFT<float> ft;
 
 	};
